Fixed minor issues and bugs in UI and Validations
* Corrected some anchors to panels and buttons in the Forms
* Changed the data grids to the same height for a more unified look
* Corrected some wrong convertions to Int32 from Decimals

+++ Session-11/Session-11/EngineerF.cs
@@ -78,7 +78,7 @@ namespace Session_11
                 MessageBox.Show("Please fill the empty fields", "Warning");
                 return;
             }
-            if (Convert.ToInt32(Ctrlsallarypermonth.EditValue.ToString()) < Ctrlsallarypermonth.Properties.MinValue)
+            if (Convert.ToDecimal(Ctrlsallarypermonth.EditValue.ToString()) < Ctrlsallarypermonth.Properties.MinValue)
             {
                 MessageBox.Show("Sallary can't be a negative number", "Warning");
                 return;
